Changed scaling of floating points to use the size of a short integer.

+++ cpp_rfnoc/TuneFilterDecimate.cpp
@@ -659,7 +659,7 @@ bool TuneFilterDecimate_i::configureFD(bool sriChanged)
     std::vector<int> longKaiserTaps(estimatedFilterLength);
 
     for (size_t i = 0; i < estimatedFilterLength; ++i) {
-        longKaiserTaps[i] = (pow(2, sizeof(int) * 8) - 1) * kaiserTaps[i];
+        longKaiserTaps[i] = (pow(2, sizeof(short int) * 8) - 1) * kaiserTaps[i];
     }
 
     LOG_DEBUG(TuneFilterDecimate_i, "Setting filter taps on RF-NoC block...");
